move websocekt setup code to lower in the stack
